third_party/WebKit/LayoutTests/external/wpt/html/semantics/embedded-content/media-elements/interfaces/TextTrack/mode-expected.txt

CONSOLE WARNING: line 10: The provided value '1' is not a valid enum value of type TextTrackMode.
CONSOLE WARNING: line 12: The provided value '' is not a valid enum value of type TextTrackMode.
CONSOLE WARNING: line 14: The provided value 'null' is not a valid enum value of type TextTrackMode.
CONSOLE WARNING: line 16: The provided value 'undefined' is not a valid enum value of type TextTrackMode.
CONSOLE WARNING: line 20: The provided value 'DISABLED' is not a valid enum value of type TextTrackMode.
CONSOLE WARNING: line 22: The provided value 'dİsabled' is not a valid enum value of type TextTrackMode.
CONSOLE WARNING: line 24: The provided value 'dısabled' is not a valid enum value of type TextTrackMode.
CONSOLE WARNING: line 26: The provided value 'disabled ' is not a valid enum value of type TextTrackMode.
CONSOLE WARNING: line 28: The provided value ' disabled' is not a valid enum value of type TextTrackMode.
CONSOLE WARNING: line 30: The provided value '[object Object]' is not a valid enum value of type TextTrackMode.
CONSOLE WARNING: line 32: The provided value 'HIDDEN' is not a valid enum value of type TextTrackMode.
CONSOLE WARNING: line 34: The provided value 'hİdden' is not a valid enum value of type TextTrackMode.
CONSOLE WARNING: line 36: The provided value 'hıdden' is not a valid enum value of type TextTrackMode.
This is a testharness.js-based test.
PASS TextTrack.mode, wrong value 
PASS TextTrack.mode, correct value 
Harness: the test ran to completion.
